ENG-13633: Initial discovery and classification implementation (#51)
Initial implementation of Dmap's discovery and classification feature.

The main entrypoint is the "dmap" CLI (see the cmd directory).

Note that some of the details like command name, parameters, etc. are subject to change until the first stable version is released.

Additionally, most of the code that powers the CLI has been added as public packages to the main module, enhancing the API of the existing Dmap library. Users can use these packages to implement their own discovery and classification tooling if desired. There are two new top-level packages added to the public API:
  * classification - provides an API to perform data classification on arbitrary string data.
  * sql - provides an API to introspect, sample, and scan (which is introspect + sample + classify) SQL data repositories.

A new RepoScanner interface was also added to the scan package.